refactored .correlate method
